Move account related structs to separate file
