+++ finality-aleph/src/idx_to_account.rs
@@ -1,0 +1,83 @@
+use std::{marker::PhantomData, sync::Arc};
+
+use primitives::{AccountId, AlephSessionApi, AuraId, BlockHash, BlockNumber};
+use sc_client_api::Backend;
+use sp_consensus_aura::AuraApi;
+use sp_runtime::traits::{Block, Header};
+
+use crate::{
+    abft::NodeIndex,
+    session::{SessionBoundaryInfo, SessionId},
+    session_map::{AuthorityProvider, AuthorityProviderImpl},
+    ClientForAleph,
+};
+
+pub trait ValidatorIndexToAccountIdConverter {
+    fn account(&self, session: SessionId, validator_index: NodeIndex) -> Option<AccountId>;
+}
+
+pub struct ValidatorIndexToAccountIdConverterImpl<C, B, BE>
+where
+    C: ClientForAleph<B, BE> + Send + Sync + 'static,
+    C::Api: crate::aleph_primitives::AlephSessionApi<B> + AuraApi<B, AuraId>,
+    B: Block<Hash = BlockHash>,
+    BE: Backend<B> + 'static,
+{
+    client: Arc<C>,
+    session_boundary_info: SessionBoundaryInfo,
+    authority_provider: AuthorityProviderImpl<C, B, BE>,
+    _phantom: PhantomData<(B, BE)>,
+}
+
+impl<C, B, BE> ValidatorIndexToAccountIdConverterImpl<C, B, BE>
+where
+    C: ClientForAleph<B, BE> + Send + Sync + 'static,
+    C::Api: crate::aleph_primitives::AlephSessionApi<B> + AuraApi<B, AuraId>,
+    B: Block<Hash = BlockHash>,
+    B::Header: Header<Number = BlockNumber>,
+    BE: Backend<B> + 'static,
+{
+    pub fn new(client: Arc<C>, session_boundary_info: SessionBoundaryInfo) -> Self {
+        Self {
+            client: client.clone(),
+            session_boundary_info,
+            authority_provider: AuthorityProviderImpl::new(client),
+            _phantom: PhantomData,
+        }
+    }
+}
+
+impl<C, B, BE> ValidatorIndexToAccountIdConverter
+    for ValidatorIndexToAccountIdConverterImpl<C, B, BE>
+where
+    C: ClientForAleph<B, BE> + Send + Sync + 'static,
+    C::Api: crate::aleph_primitives::AlephSessionApi<B> + AuraApi<B, AuraId>,
+    B: Block<Hash = BlockHash>,
+    B::Header: Header<Number = BlockNumber>,
+    BE: Backend<B> + 'static,
+{
+    fn account(&self, session: SessionId, validator_index: NodeIndex) -> Option<AccountId> {
+        let block_number = self
+            .session_boundary_info
+            .boundaries_for_session(session)
+            .first_block();
+        let block_hash = self.client.block_hash(block_number).ok()??;
+
+        let authority_data = self.authority_provider.authority_data(block_number)?;
+        let aleph_key = authority_data.authorities()[validator_index.0].clone();
+        self.client
+            .runtime_api()
+            .key_owner(block_hash, aleph_key)
+            .ok()?
+    }
+}
+
+#[cfg(test)]
+pub struct MockConverter;
+
+#[cfg(test)]
+impl ValidatorIndexToAccountIdConverter for MockConverter {
+    fn account(&self, _: SessionId, _: NodeIndex) -> Option<AccountId> {
+        None
+    }
+}
